Restructure some of the components and controller logic
